requirements.txt
streamlit
groq
python-dotenv
sentence-transformers
PyMuPDF
pandas
plotly
fpdf
torch